fix: Binomial: Changed all signatures to long to take care of edge test

+++ Binomial.java
@@ -11,7 +11,7 @@ public class Binomial {
 	}
 
 	// Computes the Binomial function, basic version.
-	public static int binomial1(int n, int k) { 
+	public static long binomial1(int n, int k) {
 		//// Repplace the following comment with your code
         if (k > n) return 0;
         if (k == 0 || n == 0) return 1;
@@ -19,12 +19,12 @@ public class Binomial {
 	 }
 	
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
+	public static long binomial(int n, int k) {
 		//// This function creates a 2D array, say memo, 
 		//// and then initializes all its elements to -1.
 		//// It then calls binomial(n, k, memo), which does all the heavy lifiting.
 		//// Replace the following statement with your code.
-        int[][] memoArray = new int[n+1][k+1];
+        long[][] memoArray = new long[n+1][k+1];
         for (int i = 0; i < memoArray.length; i++) {
             for (int j = 0; j < memoArray[i].length; j++) {
                 memoArray[i][j] = -1;
@@ -33,7 +33,7 @@ public class Binomial {
         return binomial(n, k, memoArray);
 	}
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, long[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
